requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==0.21.0
Flask-SocketIO==5.3.6